requirements.txt
Django==1.9.7
django-admin-sortable==2.0.16
